requirements.txt
SpeechRecognition==3.10.0
pyaudio==0.2.14
pyttsx3==2.90
playwright==1.40.0
rumps==0.4.0
python-dotenv==1.0.0
openai-whisper>=20250625
torch>=2.0.0
numpy>=1.20.0
noisereduce>=2.0.0
colorama>=0.4.6
